had issues with sanitize annotation. seems the function is best served as a Genome class function anyway...
