+++ genomepy/utils.py
@@ -352,149 +352,3 @@ def get_file_info(fname):
         fname = fname[:-3]
     split = os.path.splitext(fname)
     return split[1], gz
-
-
-def sanitize_annotation(genome, gtf_file=None, sizes_file=None, out_dir=None):
-    """
-    Matches the toplevel sequence names in annotation.gtf to those in genome.fa.
-
-    The fasta and gtf formats dictate the 1st field in the genome header and gtf should match.
-    In some cases the genome.fa has multiple names per header, the 1st field not matching those in the gtf.
-    If this occurs a conversion table can be made to rename the sequence names in either file.
-
-    This script changes the names in the gtf to match those in the genome.fa, if needed.
-
-    genome: class object
-        Genome class object (contains genome name and file paths)
-
-    sizes_file: str , optional
-        Path to the genome sizes file
-
-    gtf_file: str , optional
-        Path to the (gzipped) annotation.gtf file
-
-    out_dir: str , optional
-        Location of the genome (and the README.txt)
-    """
-    # set default file paths if none are given
-    if out_dir is None:
-        out_dir = os.path.dirname(genome.filename)
-    if gtf_file is None:
-        gtf_file = os.path.join(out_dir, genome.name + ".annotation.gtf.gz")
-    if sizes_file is None:
-        sizes_file = os.path.join(out_dir, genome.name + ".fa.sizes")
-    readme = os.path.join(out_dir, "README.txt")
-
-    # unzip gtf if zipped and return up-to-date name
-    if gtf_file.endswith(".gz"):
-        sp.check_call(f"gunzip -f {gtf_file}", shell=True)
-        gtf_file = gtf_file[:-3]
-
-    # Check if genome and annotation have matching chromosome/scaffold names
-    with open(gtf_file, "r") as gtf:
-        for line in gtf:
-            if not line.startswith("#"):
-                gtf_id = line.split("\t")[0]
-                break
-
-    with open(sizes_file, "r") as sizes:
-        for line in sizes:
-            fa_id = line.split("\t")[0]
-            if fa_id == gtf_id:
-                sp.check_call(f"gzip -f {gtf_file}", shell=True)
-
-                metadata, lines = read_readme(readme)
-                metadata["sanitized annotation"] = "not required"
-                write_readme(readme, metadata, lines)
-                return
-
-    # generate a gtf with matching scaffold/chromosome IDs
-    sys.stderr.write(
-        "\nGenome and annotation do not have matching sequence names! "
-        "Creating matching annotation files...\n"
-    )
-
-    # unzip genome if zipped and return up-to-date genome name
-    bgzip, genome_file = bgunzip_and_name(genome)
-
-    # determine which element in the fasta header contains the
-    # location identifiers used in the annotation.gtf
-    header = []
-    with open(genome_file, "r") as fa:
-        for line in fa:
-            if line.startswith(">"):
-                header = line.strip(">\n").split(" ")
-                break
-
-    with open(gtf_file, "r") as gtf:
-        for line in gtf:
-            if not line.startswith("#"):
-                loc_id = line.strip().split("\t")[0]
-                try:
-                    element = header.index(loc_id)
-                    break
-                except ValueError:
-                    continue
-        else:
-            # re-zip genome if it was unzipped earlier
-            bgrezip(bgzip, genome_file)
-
-            metadata, lines = read_readme(readme)
-            metadata["sanitized annotation"] = "not possible"
-            write_readme(readme, metadata, lines)
-
-            sys.stderr.write(
-                "\nWARNING: Cannot correct annotation files automatically!\n"
-                + "Leaving original version in place.\n"
-            )
-            return
-
-    # build a conversion table
-    ids = {}
-    with open(genome_file, "r") as fa:
-        for line in fa:
-            if line.startswith(">"):
-                line = line.strip(">\n").split(" ")
-                if line[element] not in ids.keys():
-                    ids.update({line[element]: line[0]})
-
-    # re-zip genome if it was unzipped earlier
-    bgrezip(bgzip, genome_file)
-
-    with TemporaryDirectory(dir=out_dir) as tmpdir:
-        # generate a gtf with updated location identifiers from the conversion table
-        new_gtf_file = os.path.join(tmpdir, os.path.basename(gtf_file))
-        missing_contigs = []
-        with open(gtf_file, "r") as oldgtf, open(new_gtf_file, "w") as newgtf:
-            for line in oldgtf:
-                line = line.split("\t")
-                try:
-                    line[0] = ids[line[0]]
-                except KeyError:
-                    missing_contigs.append(line[0])
-                line = "\t".join(line)
-                newgtf.write(line)
-
-        if len(missing_contigs) > 0:
-            sys.stderr.write(
-                "\nWARNING: annotation contains contigs not present in the genome!\n"
-                "The following contigs were not found: "
-                f"{', '.join(set(missing_contigs))}.\n"
-                "These have been kept as-is. Other contigs were corrected!\n"
-            )
-
-        # generate a bed from the gtf
-        cmd = "gtfToGenePred {0} /dev/stdout | genePredToBed /dev/stdin {1}"
-        new_bed_file = new_gtf_file.replace("gtf", "bed")
-        sp.check_call(cmd.format(new_gtf_file, new_bed_file), shell=True)
-
-        # overwrite old annotation files (and gzip)
-        for src, dst in zip(
-            [new_gtf_file, new_bed_file], [gtf_file, gtf_file.replace("gtf", "bed")]
-        ):
-            sp.check_call(f"mv {src} {dst} && gzip -f {dst}", shell=True)
-
-    metadata, lines = read_readme(readme)
-    metadata["sanitized annotation"] = "yes"
-    write_readme(readme, metadata, lines)
-    return
