+++ tests/04_genome_test.py
@@ -1,6 +1,7 @@
 import genomepy
 import os
 import pytest
+import subprocess as sp
 
 
 # to ignore file changes
@@ -25,8 +26,8 @@ def test_genome__init__(genome="tests/data/small_genome.fa.gz"):
     assert os.path.exists(g.gaps_file)
     assert isinstance(g.sizes, dict)
     assert isinstance(g.gaps, dict)
-    assert g.annotation_gtf_file is None
-    assert g.annotation_bed_file is None
+    assert "annotation_gtf_file" in dir(g)
+    assert "annotation_bed_file" in dir(g)
     assert g.tax_id == g.assembly_accession == "na"
     assert isinstance(g.plugin, dict)
 
@@ -196,6 +197,30 @@ def test_get_random_sequences(genome="tests/data/small_genome.fa.gz"):
     # check that the max Ns are lower than the expected cutoff
     rs = g.get_random_sequences(n=1, chroms=chroms, outtype="string")
     assert str(g.track2fasta(rs[0])[0].seq).upper().count("N") <= length * max_n
+
+
+def test_sanitize_annotation(genome="tests/data/small_genome.fa.gz"):
+    # generate gtf file
+    gtf_file = genome[:-5] + "annotation.gtf"
+    with open(gtf_file, "w") as f:
+        f.write("# skip this line\n")
+        f.write(
+            """chr2\tvanHeeringen-lab\tgene\t2\t22\t.\t+\t.\tgene_id "vH-1"; transcript_id "vH-1.1";\n"""
+        )
+    sp.check_call(f"gzip -f {gtf_file}", shell=True)
+
+    # generate bed file
+    bed_file = gtf_file.replace("gtf", "bed.gz")
+    sp.check_call(f"touch {bed_file}", shell=True)
+
+    g = genomepy.Genome(genome)
+    g.sanitize_annotation()
+    result = open(gtf_file).read()
+    assert result.startswith("# skip this line\nchr2\tvanHeeringen-lab")
+
+    # cleanup
+    os.unlink(gtf_file)
+    os.unlink(bed_file)
 
 
 def test_delete_test_files():
